+++ gradle/dependencies.gradle
@@ -66,7 +66,8 @@ def support = [
         lifecycle: [
             common: "android.arch.lifecycle:common:${versions.archComponents}",
             compiler: "android.arch.lifecycle:compiler:${versions.archComponents}",
-            extensions: "android.arch.lifecycle:extensions:${versions.archComponents}"
+            extensions: "android.arch.lifecycle:extensions:${versions.archComponents}",
+            runtime: "android.arch.lifecycle:runtime:1.0.0"
         ]
     ]
 ]
