+++ settings.gradle
@@ -17,6 +17,7 @@
 rootProject.name = 'autodispose-root'
 include ':android:autodispose-android'
 include ':android:autodispose-android-archcomponents'
+include ':android:autodispose-android-archcomponents-test'
 include ':autodispose'
 include ':autodispose-kotlin'
 include ':sample'
